refactor(brain): split agent.ts into modules named for their cluster
